dom/workers/test/serviceworkers/test_match_all.html

<!--
  Any copyright is dedicated to the Public Domain.
  http://creativecommons.org/publicdomain/zero/1.0/
-->
<!DOCTYPE HTML>
<html>
<head>
  <title>Bug 982726 - test match_all not crashing</title>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<p id="display"></p>
<div id="content" style="display: none"></div>
<pre id="test"></pre>
<script class="testbody" type="text/javascript">
  // match_all_worker will call matchAll until the worker shuts down.
  // Test passes if the browser doesn't crash on leaked promise objects.
  var registration;
  var content;
  var iframe;

  function simpleRegister() {
    return navigator.serviceWorker.register("match_all_worker.js",
                                            { scope: "./sw_clients/" })
      .then((swr) => registration = swr);
  }

  function closeAndUnregister() {
    content.removeChild(iframe);

    return registration.unregister().then(function(result) {
      ok(result, "Unregister should return true.");
    }, function(e) {
      dump("Unregistering the SW failed with " + e + "\n");
    });
  }

  function openClient() {
    var p = new Promise(function(resolve, reject) {
      window.onmessage = function(e) {
        if (e.data === "READY") {
          resolve();
        }
      }
    });

    content = document.getElementById("content");
    ok(content, "Parent exists.");

    iframe = document.createElement("iframe");
    iframe.setAttribute('src', "sw_clients/simple.html");
    content.appendChild(iframe);

    return p;
  }

  function runTest() {
    simpleRegister()
      .then(openClient)
      .then(closeAndUnregister)
      .catch(function(e) {
        ok(false, "Some test failed with error " + e);
      }).then(function() {
        ok(true, "Didn't crash on resolving matchAll promises while worker shuts down.");
        SimpleTest.finish();
      });
  }

  SimpleTest.waitForExplicitFinish();
  SpecialPowers.pushPrefEnv({"set": [
    ["dom.serviceWorkers.exemptFromPerDomainMax", true],
    ["dom.serviceWorkers.enabled", true],
    ["dom.serviceWorkers.testing.enabled", true]
  ]}, runTest);
</script>
</pre>
</body>
</html>
